Contents

Contents

Index

Specification and Thermodynamic State

Utilities

Defined Properties

Index

ThermoState.convert_unit

ThermoState.default_units

ThermoState.mass_density

ThermoState.mass_enthalpy

ThermoState.mass_entropy

ThermoState.mass_gibbs

ThermoState.mass_helmholtz

ThermoState.mass_internal_energy

ThermoState.mass_volume

ThermoState.mol_density

ThermoState.mol_enthalpy

ThermoState.mol_entropy

ThermoState.mol_gibbs

ThermoState.mol_helmholtz

ThermoState.mol_internal_energy

ThermoState.mol_volume

ThermoState.normalize_units

ThermoState.pressure

ThermoState.spec

ThermoState.state

ThermoState.temperature

ThermoState.total_enthalpy

ThermoState.total_entropy

ThermoState.total_gibbs

ThermoState.total_helmholtz

ThermoState.total_internal_energy

ThermoState.total_volume

ThermoState.@spec_str

ThermoState.@to_units

Specification and Thermodynamic State

ThermoState.spec — Function

spec(sp::AbstractSpec,val,normalize_units=true)
spec(;kwargs,normalize_units=true)

Generates a specification property. It can be called with the keyword interface:

sp = spec(t=32u"°C")
sp = spec(t=32u"°C",normalize_units = true)

Or the positional argument interface:

sp = spec(Types.Temperature(),32u"°C")
sp = spec(Types.Temperature(),32u"°C",true)

If normalize_units is set to false, then any unitful quantities are stored as-is, instead of being transformed to SI units.

You can use the @spec_str macro to call the apropiate Spec type:

sp = spec(spec"t",32u"°C")
sp = spec(spec"t",32u"°C",true)

ThermoState.state — Function

state(;normalize_units=true,check=true,kwargs...)
state(args...;check=true)
state((specs...))

Creates a ThermodynamicState from the arguments passed. If one or more of the arguments is the value VariableSpec(), the state created will be a callable, returning a complete state when evaluated.

Unitful quantities are normalized to SI units and unit-stripped. This can be disabled with normalize_units=false

When creating a thermodynamic state, the input arguments are checked for consistency with the gibbs rule. This check can be skipped with check=false

if called with a tuple of specs, all the checks are skipped, allowing defining states at compile time. this form does not allow variable specifications.

examples:

st1 = state(t=1,p=2,mass=3) #keyword interface

t0 = spec(spec"t",1)
p0 = spec(spec"p",2)
mass0 = spec(spec"mass",3)
st2 = state(t0,p0,mass0) # positional arguments interface

st3 = state((t0,p0,mass0)) #no runtime cost, not checked

st4 = state(t=1,T=2,p=3,P=4) #errors, spends a determinate amount of time checking for errors
st4 = state(t=1,T=2,p=3,P=4,check=false) #doesnt error, doenst check

Utilities

ThermoState.@to_units — Macro

@to_units f(x)

return f(x) with Unitful units. the measurement unit is determined by default_units(a::typeof(f))

#example:

julia> critical_pressure(water)
2.2064e7 Pa

ThermoState.@spec_str — Macro

@spec_str(term)

Allows an easy instantiation of specification properties, returning a specification type corresponding to the input symbol:

spec"t" = Temperature()

This property, among with the spec function, allows the creation of specification properties at compile time:

julia> Using Unitful,ThermoState,BenchmarkTools

julia> @btime spec(spec"t",232u"°C",false)
  0.001 ns (0 allocations: 0 bytes)
spec(t = 232[°C])

ThermoState.normalize_units — Function

normalize_units(val)

On normal numbers, it is the identity, but on numbers or vectors of Unitful.Quantity,it converts the unit to an equivalent SI unit and strips the unit information.

Examples

julia> normalize_units(0.0u"°C")
273.15

julia> normalize_units(273.15)
273.15

ThermoState.convert_unit — Function

convert_unit(from,to,val)

Converts an unit from the unit stored in from to the unit stored in to.

When both units are equal, it justs returns val.

If val itself is an unit, then it converts the from the unit in val to the unit in to.

Examples

julia> convert_unit(u"Pa",u"kPa",1000.0)      
1.0

julia> convert_unit(u"Pa",u"kPa",1u"atm")     
4053//40

ThermoState.default_units — Function

default_units(::Union{Type{AbstractSpec},Type{property_function}})

returns a default unit of a spec

Examples

default_units(pressure) #function name
Pa

default_units(Gibbs{MOLAR}) #spec type
J/mol

Defined Properties

PVT Properties

ThermoState.temperature — Function

temperature(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :t, :T

Compute the temperature (K), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.pressure — Function

pressure(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: `:p`, `:P`

Compute the pressure (Pa), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mol_volume — Function

mol_volume(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mol_v, :v

Compute the molar volume (m³/mol), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mass_volume — Function

mass_volume(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mass_v

Compute the specific (mass) volume (m³/kg), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.total_volume — Function

total_volume(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :total_v, :V

Compute the total volume (m³), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mol_density — Function

mol_density(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mol_rho, :rho, :mol_ρ, :ρ

Compute the molar density (mol/m³), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mass_density — Function

mass_density(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mass_rho, :mass_ρ

Compute the specific (mass) density (kg/m³), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

Energy Properties

ThermoState.mol_helmholtz — Function

mol_helmholtz(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mol_a, :a

Compute the molar helmholtz energy (J/mol), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mass_helmholtz — Function

mass_helmholtz(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mass_a

Compute the specific (mass) helmholtz energy (J/kg), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.total_helmholtz — Function

total_helmholtz(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :total_a

Compute the total helmholtz energy (J), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mol_gibbs — Function

mol_gibbs(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mol_g, :g

Compute the molar gibbs energy (J/mol), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mass_gibbs — Function

mass_gibbs(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mass_g

Compute the specific (mass) gibbs energy (J/kg), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.total_gibbs — Function

total_gibbs(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :total_g

Compute the total gibbs energy (J), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mol_enthalpy — Function

mol_enthalpy(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mol_h, :h

Compute the molar enthalpy (J/mol), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mass_enthalpy — Function

mass_enthalpy(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mass_h

Compute the specific (mass) enthalpy (J/kg), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.total_enthalpy — Function

total_enthalpy(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :total_h

Compute the total enthalpy (J), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mol_entropy — Function

mol_entropy(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mol_s, :s

Compute the molar entropy (J/(mol K)), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mass_entropy — Function

mass_entropy(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mass_s

Compute the specific (mass) entropy (J/(kg K)), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.total_entropy — Function

total_entropy(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :total_s

Compute the total entropy (J/K), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mol_internal_energy — Function

mol_internal_energy(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mol_u, :u

Compute the molar internal energy (J/mol), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mass_internal_energy — Function

mass_internal_energy(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mass_u

Compute the specific (mass) internal energy (J/kg), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.total_internal_energy — Function

total_internal_energy(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :total_u

Compute the total internal energy (J), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

Mass and Molar Properties

ThermoState.molfraction ThermoState.massfraction ThermoState.molnumber ThermoState.massnumber ThermoState.moles ThermoState.mass ThermoState.molar_mass

Other Properties

ThermoState.options ThermoState.phase ThermoState.quality ThermoState.molecular_weight ```