Contents

Contents

Index

Specification and Thermodynamic State

Utilities

Defined Properties

Index

ThermoState.convert_unit

ThermoState.default_units

ThermoState.get_spec

ThermoState.has_spec

ThermoState.mass

ThermoState.mass_density

ThermoState.mass_enthalpy

ThermoState.mass_entropy

ThermoState.mass_fraction

ThermoState.mass_gibbs

ThermoState.mass_helmholtz

ThermoState.mass_internal_energy

ThermoState.mass_number

ThermoState.mass_volume

ThermoState.mol_density

ThermoState.mol_enthalpy

ThermoState.mol_entropy

ThermoState.mol_fraction

ThermoState.mol_gibbs

ThermoState.mol_helmholtz

ThermoState.mol_internal_energy

ThermoState.mol_number

ThermoState.mol_volume

ThermoState.molar_mass

ThermoState.molecular_weight

ThermoState.moles

ThermoState.normalize_units

ThermoState.options

ThermoState.phase

ThermoState.pressure

ThermoState.quality

ThermoState.spec

ThermoState.state

ThermoState.state_type

ThermoState.temperature

ThermoState.total_enthalpy

ThermoState.total_entropy

ThermoState.total_gibbs

ThermoState.total_helmholtz

ThermoState.total_internal_energy

ThermoState.total_volume

ThermoState.@spec_str

ThermoState.@to_units

Specification and Thermodynamic State

ThermoState.spec — Function

spec(sp::AbstractSpec,val,normalize_units=true)
spec(;kwargs,normalize_units=true)

Generates a specification property. It can be called with the keyword interface:

sp = spec(t=32u"°C")
sp = spec(t=32u"°C",normalize_units = true)

Or the positional argument interface:

sp = spec(Types.Temperature(),32u"°C")
sp = spec(Types.Temperature(),32u"°C",true)

If normalize_units is set to false, then any unitful quantities are stored as-is, instead of being transformed to SI units.

You can use the @spec_str macro to call the apropiate Spec type:

sp = spec(spec"t",32u"°C")
sp = spec(spec"t",32u"°C",true)

ThermoState.state — Function

state(;normalize_units=true,check=true,kwargs...)
state(args...;check=true)
state((specs...))

Creates a ThermodynamicState from the arguments passed. If one or more of the arguments is the value VariableSpec(), the state created will be a callable, returning a complete state when evaluated.

Unitful quantities are normalized to SI units and unit-stripped. This can be disabled with normalize_units=false

When creating a thermodynamic state, the input arguments are checked for consistency with the gibbs rule. This check can be skipped with check=false

if called with a tuple of specs, all the checks are skipped, allowing defining states at compile time. this form does not allow variable specifications.

examples:

st1 = state(t=1,p=2,mass=3) #keyword interface

t0 = spec(spec"t",1)
p0 = spec(spec"p",2)
mass0 = spec(spec"mass",3)
st2 = state(t0,p0,mass0) # positional arguments interface

st3 = state((t0,p0,mass0)) #no runtime cost, not checked

st4 = state(t=1,T=2,p=3,P=4) #errors, spends a determinate amount of time checking for errors
st4 = state(t=1,T=2,p=3,P=4,check=false) #doesnt error, doenst check

ThermoState.get_spec — Function

get_spec(val::AbstractSpec,st::ThermodynamicState)
get_spec(val::Type{AbstractSpec},st::ThermodynamicState)

Returns the first specification of the type val present in the state st. returns nothing if not found.

Example

julia> a1 = state(n = [1.1,2.2,3.3],T=273.15,mol_h = 101.342)
ThermodynamicState with 3 properties:
  Molar amounts : [1.1, 2.2, 3.3][mol]
  Temperature : 273.15[K]
  Molar enthalpy : 101.342[J mol^-1]

julia> get_spec(Enthalpy,a1)
  spec(mol_h = 101.342[J mol^-1])

julia> get_spec(VolumeAmount,a1) == nothing
  true

ThermoState.has_spec — Function

has_spec(val::AbstractSpec,st::ThermodynamicState)::Bool
has_spec(val::Type{AbstractSpec},st::ThermodynamicState)::Bool

checks if the specification of the type val present in the state st.

Example

julia> a1 = state(n = [1.1,2.2,3.3],T=273.15,mol_h = 101.342)
ThermodynamicState with 3 properties:
  Molar amounts : [1.1, 2.2, 3.3][mol]
  Temperature : 273.15[K]
  Molar enthalpy : 101.342[J mol^-1]

julia> has_spec(Enthalpy,a1)
  true

julia> has_spec(VolumeAmount,a1) == nothing
  false

ThermoState.state_type — Function

state_type(x::ThermodynamicState)

Returns a sorted tuple of thermodynamic specifications of the input state, the tuple has 3 elements, corresponding to two general specifications and a last amount specification, indicating if the state is single or multicomponent.

Utilities

ThermoState.@to_units — Macro

@to_units f(x)

return f(x) with Unitful units. the measurement unit is determined by default_units(a::typeof(f))

#example:

julia> critical_pressure(water)
2.2064e7 Pa

ThermoState.@spec_str — Macro

@spec_str(term)

Allows an easy instantiation of specification properties, returning a specification type corresponding to the input symbol:

spec"t" = Temperature()

This property, among with the spec function, allows the creation of specification properties at compile time:

julia> Using Unitful,ThermoState,BenchmarkTools

julia> @btime spec(spec"t",232u"°C",false)
  0.001 ns (0 allocations: 0 bytes)
spec(t = 232[°C])

ThermoState.normalize_units — Function

normalize_units(val)

On normal numbers, it is the identity, but on numbers or vectors of Unitful.Quantity,it converts the unit to an equivalent SI unit and strips the unit information.

Examples

julia> normalize_units(0.0u"°C")
273.15

julia> normalize_units(273.15)
273.15

ThermoState.convert_unit — Function

convert_unit(from,to,val)

Converts an unit from the unit stored in from to the unit stored in to.

When both units are equal, it justs returns val.

If val itself is an unit, then it converts the from the unit in val to the unit in to.

Examples

julia> convert_unit(u"Pa",u"kPa",1000.0)      
1.0

julia> convert_unit(u"Pa",u"kPa",1u"atm")     
4053//40

ThermoState.default_units — Function

default_units(::Union{Type{AbstractSpec},Type{property_function}})

returns a default unit of a spec

Examples

default_units(pressure) #function name
Pa

default_units(Gibbs{MOLAR}) #spec type
J/mol

Defined Properties

The following properties are accepted by an state and have an accesor function:

Property	Units	Accessor function	Keywords for spec and state

Molar volume	m3/mol	mol_volume	v, mol_v

Mass volume	m3/kg	mass_volume	mass_v

Total volume	m3	total_volume	total_v, V

Molar density	mol/m3	mol_density	ρ, mol_ρ, rho, mol_rho

Mass density	mol/kg	mass_density	mass_ρ, mass_rho

Temperature	K	temperature	t, T

Pressure	Pa	pressure	p, P

Molar enthalpy	J/mol	mol_enthalpy	h, mol_h

Mass enthalpy	J/kg	mass_enthalpy	mass_h

Total enthalpy	J	total_enthalpy	total_h

Molar Gibbs	J/mol	mol_gibbs	g, mol_g

Mass Gibbs	J/kg	mass_gibbs	mass_g

Total Gibbs	J	total_gibbs	total_g

Molar Helmholtz	J/mol	mol_helmholtz	a, mol_a

Mass Helmholtz	J/kg	mass_helmholtz	mass_a

Total Helmholtz	J	total_helmholtz	total_a

Molar Internal Energy	J/mol	mol_internal_energy	a, mol_a

Mass Internal Energy	J/kg	mass_internal_energy	mass_a

Total Internal Energy	J	total_internal_energy	total_a

Molar entropy	J/(mol K)	mol_entropy	s, mol_s

Mass entropy	J/(kg K)	mass_entropy	mass_s

Total entropy	J/K	total_entropy	total_s

Amount of Moles	mol	moles	moles

Amount of Mass	kg	mass	mass

Molar Fraction	no units	mol_fraction	xn

Mass Fraction	no units	mass_fraction	xm

Molar Numbers	mol	mol_number	n

Mass Numbers	kg	mass_number	m

Molar Vapor Fraction	no units	mol_vapor_fraction	vfrac

Mass Vapor Fraction	no units	mass_vapor_fraction	quality

Current Phase	no units	phase	phase, sat

Aditional Options	no units	options	options

PVT Properties

ThermoState.temperature — Function

temperature(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :t, :T

Compute the temperature (K), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.pressure — Function

pressure(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: `:p`, `:P`

Compute the pressure (Pa), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mol_volume — Function

mol_volume(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mol_v, :v

Compute the molar volume (m³/mol), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mass_volume — Function

mass_volume(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mass_v

Compute the specific (mass) volume (m³/kg), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.total_volume — Function

total_volume(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :total_v, :V

Compute the total volume (m³), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mol_density — Function

mol_density(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mol_rho, :rho, :mol_ρ, :ρ

Compute the molar density (mol/m³), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mass_density — Function

mass_density(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mass_rho, :mass_ρ

Compute the specific (mass) density (kg/m³), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

Energy Properties

ThermoState.mol_helmholtz — Function

mol_helmholtz(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mol_a, :a

Compute the molar helmholtz energy (J/mol), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mass_helmholtz — Function

mass_helmholtz(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mass_a

Compute the specific (mass) helmholtz energy (J/kg), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.total_helmholtz — Function

total_helmholtz(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :total_a

Compute the total helmholtz energy (J), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mol_gibbs — Function

mol_gibbs(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mol_g, :g

Compute the molar gibbs energy (J/mol), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mass_gibbs — Function

mass_gibbs(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mass_g

Compute the specific (mass) gibbs energy (J/kg), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.total_gibbs — Function

total_gibbs(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :total_g

Compute the total gibbs energy (J), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mol_enthalpy — Function

mol_enthalpy(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mol_h, :h

Compute the molar enthalpy (J/mol), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mass_enthalpy — Function

mass_enthalpy(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mass_h

Compute the specific (mass) enthalpy (J/kg), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.total_enthalpy — Function

total_enthalpy(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :total_h

Compute the total enthalpy (J), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mol_entropy — Function

mol_entropy(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mol_s, :s

Compute the molar entropy (J/(mol K)), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mass_entropy — Function

mass_entropy(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mass_s

Compute the specific (mass) entropy (J/(kg K)), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.total_entropy — Function

total_entropy(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :total_s

Compute the total entropy (J/K), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mol_internal_energy — Function

mol_internal_energy(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mol_u, :u

Compute the molar internal energy (J/mol), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mass_internal_energy — Function

mass_internal_energy(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mass_u

Compute the specific (mass) internal energy (J/kg), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.total_internal_energy — Function

total_internal_energy(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :total_u

Compute the total internal energy (J), from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

Mass and Molar Properties

ThermoState.mol_fraction — Function

mol_fraction(model,state::ThermodynamicState,unit,[options...])::AbstractVector

Keyword symbols: :xn

Compute the molar fraction of each present compound present in the provided system, from or using the specifications as base.

The unit argument is not used here. but its present for consistency.

ThermoState.mass_fraction — Function

mass_fraction(model,state::ThermodynamicState,unit,[options...])::AbstractVector

Keyword symbols: :xm

Compute the mass fraction of each present compound present in the provided system, from or using the specifications as base.

The unit argument is not used here. but its present for consistency.

ThermoState.mol_number — Function

mol_number(model,state::ThermodynamicState,unit,[options...])::AbstractVector

Keyword symbols: :n

Compute the amount of moles (in mol) of each present compound present in the provided system, from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mass_number — Function

mass_number(model,state::ThermodynamicState,unit,[options...])::AbstractVector

Keyword symbols: :m

Compute the amount of mass (in kg) of each present compound present in the provided system, from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.moles — Function

moles(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :moles

Compute the total amount of moles (in mol) of the provided system, from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.mass — Function

mass(model,state::ThermodynamicState,unit,[options...])::Real

Keyword symbols: :mass

Compute the total amount of mass (in kg) of the provided system, from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

ThermoState.molar_mass — Function

molar_mass(model,state::ThermodynamicState,unit,[options...])::Real

Compute the amount of mass per unit of moles (in kg/mol) of the provided state, from or using the specifications as base.

If you need unitful units, use the macro @to_units before the function call

Other Properties

ThermoState.options — Function

options(model,state::ThermodynamicState)::NamedTuple

Keyword symbols: :options

Returns any options related to the model, stored in the state or the model.

Returns an empty named tuple otherwise.

ThermoState.phase — Function

phase(model,state::ThermodynamicState)::NamedTuple

Keyword symbols: :phase

Returns a symbol indicating the phase specified in the state or the model.

Returns :unspecified otherwise.

ThermoState.quality — Function

quality(model,state::ThermodynamicState)::NamedTuple

Keyword symbols: :quality

Returns the vapor quality stored in the state

Returns NaN otherwise.

This function is an alias for mass_vapor_fraction.

ThermoState.molecular_weight — Function

molecular_weight(model)

Keyword symbols: :mw

Returns the molecular weight stored in a model

Returns nothing if not defined.

it should return a list of the molecular weights. in g/mol. You can return a single number if the model has only one component.